♻️Refactor - 이미지 경로 변경, 더미 데이터 예외처리 변경 #44
