+++ src/constants/seminar/seminarReviewData.ts
@@ -5,32 +5,8 @@ export const SEMINAR_REVIEWS_DATA: SeminarReview[][] = [
         {
             id: 1,
             seminar_id: '',
-            author: '서희찬',
-            content:'리뷰입니다.',
-        },
-        {
-            id: 2,
-            seminar_id: '',
-            author: '서희찬',
-            content:'리뷰입니다.',
-        },
-        {
-            id: 3,
-            seminar_id: '',
-            author: '서희찬',
-            content:'리뷰입니다.',
-        },
-        {
-            id: 4,
-            seminar_id: '',
-            author: '서희찬',
-            content:'리뷰입니다.',
-        },
-        {
-            id: 5,
-            seminar_id: '',
-            author: '서희찬',
-            content:'리뷰입니다.',
+            author: '',
+            content:'',
         },
     ],
     [
